Restructure
Moved all user related classes to package com.deiksoftdev.automatagame.user
and global RestExceptionHandler to com.deiksoftdev.automatagame.rest.exception
