+++ src/main/java/com/deiksoftdev/automatagame/rest/exception/RestExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.deiksoftdev.automatagame.controller;
+package com.deiksoftdev.automatagame.rest.exception;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpHeaders;
@@ -6,10 +6,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-@ControllerAdvice
+import com.deiksoftdev.automatagame.user.UserNotFoundException;
+
+@ControllerAdvice(annotations = RestController.class)
 public class RestExceptionHandler extends ResponseEntityExceptionHandler {
  
     @ExceptionHandler({ UserNotFoundException.class })
