+++ src/main/java/com/deiksoftdev/automatagame/user/UserRestController.java
@@ -1,4 +1,4 @@
-package com.deiksoftdev.automatagame.controller;
+package com.deiksoftdev.automatagame.user;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,9 +9,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-
-import com.deiksoftdev.automatagame.model.User;
-import com.deiksoftdev.automatagame.repository.UserRepository;
 
 @RestController
 @RequestMapping(value = UserRestController.USER_API)
