+++ src/test/java/com/deiksoftdev/automatagame/user/UserRestControllerTests.java
@@ -1,4 +1,4 @@
-package com.deiksoftdev.automatagame;
+package com.deiksoftdev.automatagame.user;
 
 import static org.junit.Assert.assertEquals;
 
@@ -11,8 +11,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import com.deiksoftdev.automatagame.controller.UserRestController;
-import com.deiksoftdev.automatagame.model.User;
+import com.deiksoftdev.automatagame.user.User;
+import com.deiksoftdev.automatagame.user.UserRestController;
 
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
